Add react navigation and create seperate screens
