+++ apps/test-suite/SelectScreen.js
@@ -7,10 +7,11 @@ import {
   View,
   Text,
   Platform,
+  Alert,
   Button,
-  StatusBar,
   PixelRatio,
 } from 'react-native';
+import getTestModules from './Modules';
 
 class ListItem extends React.Component {
   onPress = () => {
@@ -44,7 +45,11 @@ class ListItem extends React.Component {
 export default class SelectScreen extends React.PureComponent {
   constructor(props) {
     super(props);
-    const initialValues = this.props.data.map(item => [item.name, false]);
+    this.allModules = getTestModules();
+    this.tests = this.allModules.map(module => {
+      return { name: module.name };
+    });
+    const initialValues = this.tests.map(item => [item.name, false]);
 
     this.state = {
       selected: new Map(initialValues),
@@ -53,6 +58,10 @@ export default class SelectScreen extends React.PureComponent {
     this._selectAll = this._selectAll.bind(this);
     this._getSelected = this._getSelected.bind(this);
   }
+
+  static navigationOptions = {
+    title: 'Test Suite',
+  };
 
   _keyExtractor = item => item.name;
 
@@ -85,13 +94,25 @@ export default class SelectScreen extends React.PureComponent {
 
   _getSelected = () => {
     const selected = this.state.selected;
-    let selectedTests = [];
+    const set = new Set();
     for (const test of selected.keys()) {
       if (selected.get(test)) {
-        selectedTests.push(test);
+        set.add(test);
       }
     }
-    return new Set(selectedTests);
+    const selectedModules = this.allModules.filter(m => set.has(m.name));
+    return selectedModules;
+  };
+
+  _navigateToTests = () => {
+    const selected = this._getSelected();
+    if (selected.size === 0) {
+      Alert.alert('Cannot Run Tests', 'You must select at least one test to run.');
+    } else {
+      this.props.navigation.navigate('RunTests', {
+        selected,
+      });
+    }
   };
 
   render() {
@@ -99,18 +120,17 @@ export default class SelectScreen extends React.PureComponent {
     return (
       <View style={styles.mainContainer}>
         <FlatList
-          data={this.props.data}
+          data={this.tests}
           extraData={this.state}
           keyExtractor={this._keyExtractor}
           renderItem={this._renderItem}
-          style={styles.flatList}
         />
         <View style={styles.buttonRow}>
           <View style={styles.buttonContainer}>
             <Button title={buttonTitle} onPress={this._selectAll} />
           </View>
           <View style={styles.buttonContainer}>
-            <Button title="Run Tests" onPress={() => this.props.runTests(this._getSelected())} />
+            <Button title="Run Tests" onPress={() => this._navigateToTests()} />
           </View>
         </View>
       </View>
@@ -121,9 +141,6 @@ export default class SelectScreen extends React.PureComponent {
 const styles = StyleSheet.create({
   mainContainer: {
     flex: 1,
-  },
-  flatList: {
-    marginTop: StatusBar.currentHeight,
   },
   listItem: {
     paddingHorizontal: 10,
